requirements.txt

allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest==8.3.2
requests==2.32.3
selenium==4.23.1

